refactor: move activity view types to a layer-neutral module
Server (SSR) code type-imported the activity view-model types (Repo, Language, YearCount, FilterState, ActivityState) from the Vue client composable useActivityApi.ts, which pulls in vue and astro:actions. Relocating those pure data types into src/activity/types.ts lets server code depend on a layer-neutral module and stops it from reaching into client code. Pure type relocation with no behavior change.
