+++ src/components/activity/ActivityTimeline.vue
@@ -1,10 +1,7 @@
 <script setup lang="ts">
 import { computed, onMounted, onUnmounted, ref } from "vue";
-import {
-  useActivityApi,
-  type Repo,
-  type YearCount,
-} from "./composables/useActivityApi";
+import { useActivityApi } from "./composables/useActivityApi";
+import type { Repo, YearCount } from "@/activity/types";
 import FilterControls from "./FilterControls.vue";
 import LanguageBar from "./LanguageBar.vue";
 import RepoCard from "./RepoCard.vue";
