+++ src/components/activity/TimelineRail.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed, ref } from "vue";
-import type { YearCount } from "./composables/useActivityApi";
+import type { YearCount } from "@/activity/types";
 
 const MAX_VISIBLE = 7;
