+++ src/pages/activity/code/_markdown.ts
@@ -1,4 +1,4 @@
-import type { Repo } from "@/components/activity/composables/useActivityApi";
+import type { Repo } from "@/activity/types";
 
 export function formatReposMarkdown(
   repos: Repo[],
